Extract PeerLinkChannel into channel.py
Third PR in the peer_link split arc (follows #769 + #770).
Moves the PeerLinkChannel dataclass byte-for-byte from
peer_link/__init__.py into peer_link/channel.py.

External callers (peer_link_client/_submit.py,
peer_link_client/client.py) keep using
"from .peer_link import PeerLinkChannel" / equivalent via
the re-export at the bottom of __init__.py.

The re-export goes at the BOTTOM of __init__.py (not next to
the other imports at the top) because channel.py has lazy
imports for _send_bytes_safely and parse_app_frame — both
still live in __init__.py until PRs 4 and 6 extract them.
Loading channel.py before those definitions exist on the
package would fail; loading after is the natural order.

The two cross-module helpers are imported lazily inside the
channel methods to keep this PR strictly mechanical. Once
_send_bytes_safely lands in plumbing.py and parse_app_frame
in session.py, the lazy imports become regular module-level
sibling imports.

__init__.py: 977 -> 893 lines (-84). channel.py: 119 lines.
aiohttp top-level import dropped from __init__.py (was only
used by the channel).

All 3389 non-e2e tests pass; ruff + mypy clean.
